refactor: update version location
